tools/requirements.txt
pyserial>=3.5
bleak>=0.20.0
